LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrack/removeCue-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.

PASS TextTrack.removeCue(), two elementless tracks 
FAIL TextTrack.removeCue(), cue from track element assert_unreached: got error event Reached unreachable code
